+++ segeval/agreement/kappa_test.py
@@ -22,8 +22,8 @@ class TestKappa(unittest.TestCase):
         Test Kappa upon Group 5 of Kazantseva (2012) data.
         '''
         data = KAZANTSEVA2012_G5
-        self.assertEqual(fleiss_kappa_linear(data),
-                         Decimal('0.2374030241245084051985691283'))
+        self.assertAlmostEqual(fleiss_kappa_linear(data),
+                               Decimal('0.2421099324579046442791865267'))
 
     def test_fleiss_kappa_linear_g5_ch1(self):
         '''
@@ -31,16 +31,16 @@ class TestKappa(unittest.TestCase):
         '''
         data = KAZANTSEVA2012_G5['ch1']
         data = {'ch1': data}
-        self.assertEqual(fleiss_kappa_linear(data),
-                         Decimal('0.1940298507462686567164179104'))
+        self.assertAlmostEqual(fleiss_kappa_linear(data),
+                               Decimal('0.2258064516129032258064516129'))
 
     def test_fleiss_kappa_linear_g2(self):
         '''
         Test Kappa upon Group 2 of Kazantseva (2012) data.
         '''
         data = KAZANTSEVA2012_G2
-        self.assertEqual(fleiss_kappa_linear(data),
-                         Decimal('0.4068521352341765453201908783'))
+        self.assertAlmostEqual(fleiss_kappa_linear(data),
+                               Decimal('0.4106841873722731313872632664'))
 
     def test_fleiss_kappa_linear_g2_ch2(self):
         '''
@@ -48,16 +48,16 @@ class TestKappa(unittest.TestCase):
         '''
         data = KAZANTSEVA2012_G2['ch2']
         data = {'ch2': data}
-        self.assertEqual(fleiss_kappa_linear(data),
-                         Decimal('0.5197205281147376181221097781'))
+        self.assertAlmostEqual(fleiss_kappa_linear(data),
+                               Decimal('0.5339895791843801565560911374'))
 
     def test_fleiss_kappa_disagree(self):
         '''
         Test Kappa upon a hypothetical dataset containing large disagreement.
         '''
         data = LARGE_DISAGREEMENT
-        self.assertEqual(fleiss_kappa_linear(data),
-                         Decimal('-0.05952156715012243360331512524'))
+        self.assertAlmostEqual(fleiss_kappa_linear(data),
+                               Decimal('0'))
 
     def test_fleiss_kappa(self):
         '''
@@ -67,15 +67,15 @@ class TestKappa(unittest.TestCase):
                         'c2': [2, 1, 7, 2, 1]}}
         kappa1 = fleiss_kappa_linear(data1)
         kappa1f = fleiss_kappa_linear(data1)
-        self.assertEqual(kappa1,
-                         Decimal('0.7096774193548387096774193548'))
+        self.assertAlmostEqual(kappa1,
+                               Decimal('0.7272727272727272727272727273'))
         self.assertEqual(kappa1, kappa1f)
         data2 = {'i1': {'c1': [2, 8, 2, 1],
                         'c2': [11, 2]}}
         kappa2 = fleiss_kappa_linear(data2)
         kappa2f = fleiss_kappa_linear(data2)
-        self.assertEqual(kappa2,
-                         Decimal('0.1176470588235294117647058823'))
+        self.assertAlmostEqual(kappa2,
+                               Decimal('0.1489361702127659574468085107'))
         self.assertEqual(kappa2, kappa2f)
         self.assertTrue(kappa2 < kappa1)
 
@@ -92,7 +92,7 @@ class TestKappa(unittest.TestCase):
         Test multiple boundaries.
         '''
         value = fleiss_kappa_linear(MULTIPLE_BOUNDARY_TYPES)
-        self.assertEqual(value, Decimal('0.3333333333333333333333333333'))
+        self.assertAlmostEqual(value, Decimal('0.4718309859154929577464788733'))
 
     def test_exception_coders(self):
         '''