+++ segeval/agreement/pi_test.py
@@ -21,8 +21,8 @@ class TestPi(unittest.TestCase):
         '''
         Test Pi upon Group 5 of Kazantseva (2012) data.
         '''
-        self.assertEqual(fleiss_pi_linear(KAZANTSEVA2012_G5),
-                         Decimal('0.2307643892167668335086819753'))
+        self.assertAlmostEqual(fleiss_pi_linear(KAZANTSEVA2012_G5),
+                               Decimal('0.2399575652197196472212284944'))
 
     def test_fliess_pi_g5_ch1(self):
         '''
@@ -30,14 +30,14 @@ class TestPi(unittest.TestCase):
         '''
         data = {'ch1': KAZANTSEVA2012_G5['ch1']}
         self.assertAlmostEqual(fleiss_pi_linear(data),
-                               Decimal('0.1906323185011709601873536298'))
+                               Decimal('0.2226720647773279352226720649'))
 
     def test_fleiss_pi_g2(self):
         '''
         Test Pi upon Group 2 of Kazantseva (2012) data.
         '''
         self.assertAlmostEqual(fleiss_pi_linear(KAZANTSEVA2012_G2),
-                               Decimal('0.4018239928733601859131343866'))
+                               Decimal('0.4083169765912929422042553174'))
 
     def test_fleiss_pi_g2_ch2(self):
         '''
@@ -46,15 +46,15 @@ class TestPi(unittest.TestCase):
         data = {'ch2': KAZANTSEVA2012_G2['ch2']}
         # Test
         self.assertAlmostEqual(fleiss_pi_linear(data),
-                               Decimal('0.5192587209302325581395348837'))
+                               Decimal('0.5335548172757475083056478405'))
 
     def test_fleiss_pi_disagree(self):
         '''
         Test Pi upon a hypothetical dataset containing large disagreement.
         '''
         data = LARGE_DISAGREEMENT
-        self.assertEqual(fleiss_pi_linear(data),
-                         Decimal('-0.5757942099675148626179719687'))
+        self.assertAlmostEqual(fleiss_pi_linear(data),
+                               Decimal('-0.3333333333333333333333333333'))
 
     def test_fleiss_pi(self):
         '''
@@ -65,15 +65,15 @@ class TestPi(unittest.TestCase):
                         'c2': [2, 1, 7, 2, 1]}}
         pi1 = fleiss_pi_linear(data1)
         pi1f = fleiss_pi_linear(data1)
-        self.assertEqual(pi1,
-                         Decimal('0.7090909090909090909090909091'))
+        self.assertAlmostEqual(pi1,
+                               Decimal('0.7267552182163187855787476281'))
         self.assertEqual(pi1, pi1f)
         data2 = {'i1': {'c1': [2, 8, 2, 1],
                         'c2': [11, 2]}}
         pi2 = fleiss_pi_linear(data2)
         pi2f = fleiss_pi_linear(data2)
-        self.assertEqual(pi2,
-                         Decimal('0.1111111111111111111111111111'))
+        self.assertAlmostEqual(pi2,
+                               Decimal('0.1428571428571428571428571429'))
         self.assertEqual(pi2, pi2f)
         self.assertTrue(pi2 < pi1)
 
@@ -91,7 +91,7 @@ class TestPi(unittest.TestCase):
         Test multiple boundaries.
         '''
         value = fleiss_pi_linear(MULTIPLE_BOUNDARY_TYPES)
-        self.assertEqual(value, Decimal('0.3333333333333333333333333333'))
+        self.assertAlmostEqual(value, Decimal('0.4666666666666666666666666667'))
 
     def test_exception_coders(self):
         '''