+++ segeval/format_test.py
@@ -4,8 +4,12 @@ Tests segmentation format utility functions.
 .. moduleauthor:: Chris Fournier <chris.m.fournier@gmail.com>
 '''
 import unittest
-from .format import (convert_positions_to_masses, convert_masses_to_positions,
-                     boundary_string_from_masses, convert_nltk_to_masses)
+from __future__ import absolute_import
+from segeval.util.test import TestCase
+from segeval.format import (convert_positions_to_masses,
+                            convert_masses_to_positions,
+                            boundary_string_from_masses,
+                            convert_nltk_to_masses)
 
 
 class TestSegeval(unittest.TestCase):